+++ CHANGELOG.md
@@ -16,6 +16,14 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Breaking
 
+## 2023-11-24
+
+- yellowstone-grpc-geyser-1.11.1+solana.1.17.6
+
+### Fixes
+
+- geyser: reconstruct blocks with zero entries ([#245](https://github.com/rpcpool/yellowstone-grpc/pull/245))
+
 ## 2023-11-21
 
 - yellowstone-grpc-client-1.12.0+solana.1.17.6